main/dbus: fix up services, drop elogind dep
The only practical purpose logind support has is that dbus can
check if the user is logged in on a seat, which is used for the
at-console policy stuff in dbus. However, there are very few
things that use at-console, and it's generally deprecated and
a bad idea, so we won't be missing it.